Code work
working on finding my own way of moving an object between two positions.
Keeping the code simple.
Unfortunately position is offset from P1 and P2, so even though it moves, it's position is never actually equal to them. This is problematic for getting it to move back to P1 after reaching P2